Soil/ai-model/requirements.txt
torch==2.1.0
torchvision==0.16.0
pillow==10.1.0
numpy==1.24.3
tqdm==4.66.1
onnx==1.15.0
onnxruntime==1.16.3
scikit-learn==1.3.2
matplotlib==3.8.2